Rename StorageList to StoraceDevices. Next level will be "StorageContent".
