+++ app/src/main/java/eu/darken/sdmse/main/ui/dashboard/DashboardFragmentVM.kt
@@ -235,7 +235,7 @@ class DashboardFragmentVM @Inject constructor(
             data = data,
             progress = progress,
             onViewDetails = {
-                DashboardFragmentDirections.actionDashboardFragmentToStorageListFragment().navigate()
+                DashboardFragmentDirections.actionDashboardFragmentToStorageDevicesFragment().navigate()
             }
         )
     }
