+++ app/src/main/res/layout/analyzer_storage_content_apps_item.xml
