+++ app/src/main/res/layout/analyzer_storage_content_fragment.xml
@@ -27,7 +27,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/toolbar"
-        tools:listitem="@layout/analyzer_storage_list_item" />
+        tools:listitem="@layout/analyzer_storage_devices_item" />
 
     <eu.darken.sdmse.common.progress.ProgressOverlayView
         android:id="@+id/loading_overlay"
