+++ app/src/main/res/navigation/main_nav.xml
@@ -78,8 +78,8 @@
             android:id="@+id/action_dashboardFragment_to_schedulerManagerFragment"
             app:destination="@id/schedulerManagerFragment" />
         <action
-            android:id="@+id/action_dashboardFragment_to_storageListFragment"
-            app:destination="@id/storageListFragment" />
+            android:id="@+id/action_dashboardFragment_to_storageDevicesFragment"
+            app:destination="@id/storageDevicesFragment" />
     </fragment>
 
     <action
@@ -280,8 +280,8 @@
             app:argType="string" />
     </dialog>
     <fragment
-        android:id="@+id/storageListFragment"
-        android:name="eu.darken.sdmse.analyzer.ui.storage.StorageListFragment"
-        tools:layout="@layout/analyzer_storage_list_fragment" />
+        android:id="@+id/storageDevicesFragment"
+        android:name="eu.darken.sdmse.analyzer.ui.storage.devices.StorageDevicesFragment"
+        tools:layout="@layout/analyzer_storage_devices_fragment" />
 
 </navigation>
